+++ pyg_lib/csrc/sampler/neighbor.cpp
@@ -20,6 +20,7 @@ neighbor_sample(const at::Tensor& rowptr,
                 const std::vector<int64_t>& num_neighbors,
                 const c10::optional<at::Tensor>& time,
                 const c10::optional<at::Tensor>& seed_time,
+                const c10::optional<at::Tensor>& edge_weight,
                 bool csc,
                 bool replace,
                 bool directed,
@@ -37,8 +38,8 @@ neighbor_sample(const at::Tensor& rowptr,
   static auto op = c10::Dispatcher::singleton()
                        .findSchemaOrThrow("pyg::neighbor_sample", "")
                        .typed<decltype(neighbor_sample)>();
-  return op.call(rowptr, col, seed, num_neighbors, time, seed_time, csc,
-                 replace, directed, disjoint, temporal_strategy,
+  return op.call(rowptr, col, seed, num_neighbors, time, seed_time, edge_weight,
+                 csc, replace, directed, disjoint, temporal_strategy,
                  return_edge_id);
 }
 
@@ -94,10 +95,11 @@ hetero_neighbor_sample(
 TORCH_LIBRARY_FRAGMENT(pyg, m) {
   m.def(TORCH_SELECTIVE_SCHEMA(
       "pyg::neighbor_sample(Tensor rowptr, Tensor col, Tensor seed, int[] "
-      "num_neighbors, Tensor? time = None, Tensor? seed_time = None, bool csc "
-      "= False, bool replace = False, bool directed = True, bool disjoint = "
-      "False, str temporal_strategy = 'uniform', bool return_edge_id = True) "
-      "-> (Tensor, Tensor, Tensor, Tensor?, int[], int[])"));
+      "num_neighbors, Tensor? time = None, Tensor? seed_time = None, Tensor? "
+      "edge_weight = None, bool csc = False, bool replace = False, bool "
+      "directed = True, bool disjoint = False, str temporal_strategy = "
+      "'uniform', bool return_edge_id = True) -> "
+      "(Tensor, Tensor, Tensor, Tensor?, int[], int[])"));
   m.def(TORCH_SELECTIVE_SCHEMA(
       "pyg::hetero_neighbor_sample(str[] node_types, (str, str, str)[] "
       "edge_types, Dict(str, Tensor) rowptr_dict, Dict(str, Tensor) col_dict, "